back/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
Markdown==3.3.2
psycopg2-binary==2.9.3
pytz==2020.1
sqlparse==0.4.1
django-cors-headers==3.11.0
